[+BUGFIX] FLOW3 (Resource): Added a safeguard into the ResourceObjectConverter which detects if a file upload has previously been converted to an object. Without this check it could happen that FLOW3 tries to move the uploaded file multiple times, which doesn't work of course. [+BUGFIX] FLOW3 (Validation): Added a type safeguard to the StringLengthValidator Original-Commit-Hash: 8ddfdb724433d81adc39fd1a0a8ecf9ad4aa6f3f

+++ TYPO3.Flow/Classes/Resource/ResourceManager.php
@@ -202,6 +202,9 @@ class ResourceManager {
 			return FALSE;
 		}
 
+		if (!file_exists($uploadInfo['tmp_name'])) {
+			return FALSE;
+		}
 		$hash = sha1_file($uploadInfo['tmp_name']);
 		$finalTargetPathAndFilename = $this->persistentResourcesStorageBaseUri . $hash;
 		if (move_uploaded_file($uploadInfo['tmp_name'], $finalTargetPathAndFilename) === FALSE) {
